examples/reverse-xds/old-python-server/envoy_protos/envoy/service/discovery/v3/ads_pb2_grpc.py
# Minimal ADS gRPC stub for demo purposes
import grpc
from . import discovery_pb2

class AggregatedDiscoveryServiceStub:
    """Client stub for AggregatedDiscoveryService"""
    
    def __init__(self, channel):
        self.channel = channel
        self.StreamAggregatedResources = channel.stream_stream(
            '/envoy.service.discovery.v3.AggregatedDiscoveryService/StreamAggregatedResources',
            request_serializer=self._serialize_request,
            response_deserializer=self._deserialize_response,
        )
    
    def _serialize_request(self, request):
        """Serialize request for gRPC"""
        if hasattr(request, 'SerializeToString'):
            return request.SerializeToString()
        # For our minimal implementation
        import json
        return json.dumps(request.__dict__).encode('utf-8')
    
    def _deserialize_response(self, response_bytes):
        """Deserialize response from gRPC"""
        # For our minimal implementation
        import json
        data = json.loads(response_bytes.decode('utf-8'))
        response = discovery_pb2.DiscoveryResponse()
        response.__dict__.update(data)
        return response

class AggregatedDiscoveryServiceServicer:
    """Server servicer for AggregatedDiscoveryService"""
    
    def StreamAggregatedResources(self, request_iterator, context):
        """Handle bidirectional streaming RPC"""
        raise NotImplementedError('Method not implemented!')

def add_AggregatedDiscoveryServiceServicer_to_server(servicer, server):
    """Register the servicer with the gRPC server"""
    rpc_method_handlers = {
        'StreamAggregatedResources': grpc.stream_stream_rpc_method_handler(
            servicer.StreamAggregatedResources,
            request_deserializer=_deserialize_request,
            response_serializer=_serialize_response,
        ),
    }
    generic_handler = grpc.method_handlers_generic_handler(
        'envoy.service.discovery.v3.AggregatedDiscoveryService', 
        rpc_method_handlers
    )
    server.add_generic_rpc_handlers((generic_handler,))

def _deserialize_request(request_bytes):
    """Deserialize request from gRPC"""
    # For our minimal implementation
    import json
    data = json.loads(request_bytes.decode('utf-8'))
    request = discovery_pb2.DiscoveryRequest()
    request.__dict__.update(data)
    return request

def _serialize_response(response):
    """Serialize response for gRPC"""
    if hasattr(response, 'SerializeToString'):
        return response.SerializeToString()
    # For our minimal implementation
    import json
    return json.dumps(response.__dict__).encode('utf-8')